pyproject.toml
[project]
name = "dynamic-skills"
version = "0.1.0"
description = "Dynamic skill learning for Claude Code conversations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "claude-code-sdk",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.scripts]
skills-agent = "dynamic_skills.cli:agent_main"
skills-observer = "dynamic_skills.cli:observer_main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dynamic_skills"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
